+++ src/hunt/testing/__init__.py
@@ -1,4 +1,4 @@
-from hunt.testing.fakes import EventFake, MailFake, QueueFake, freeze_time
+from hunt.testing.fakes import EventFake, MailFake, NotificationFake, QueueFake, freeze_time
 from hunt.testing.test_case import DatabaseTransactions, HuntTestCase, RefreshDatabase
 
 __all__ = [
@@ -6,6 +6,7 @@ __all__ = [
     "EventFake",
     "HuntTestCase",
     "MailFake",
+    "NotificationFake",
     "QueueFake",
     "RefreshDatabase",
     "freeze_time",